requirements.txt
googlemaps>=4.10.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
rapidfuzz>=3.5.0
streamlit>=1.29.0
python-dotenv>=1.0.0
pandas>=2.1.0
requests>=2.31.0
python-pptx>=1.0.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
gTTS>=2.5.0
